+++ yggdrash-core/src/test/java/io/yggdrash/core/net/YggdrashNodeTest.java
@@ -9,13 +9,13 @@ public class YggdrashNodeTest {
 
     @Test
     public void bootstrapping() {
-        YggDrashTestNode node1 = new YggDrashTestNode("ynode://75bff16c@127.0.0.1:32918");
+        YggdrashTestNode node1 = new YggdrashTestNode("ynode://75bff16c@127.0.0.1:32918");
         node1.bootstrapping();
     }
 
-    private class YggDrashTestNode extends YggdrashNode {
+    private class YggdrashTestNode extends Node {
 
-        YggDrashTestNode(String ynodeUri) {
+        YggdrashTestNode(String ynodeUri) {
             this.discovery = new KademliaDiscoveryMock(Peer.valueOf(ynodeUri));
             List<String> seedList = Collections.singletonList("ynode://75bff16c@127.0.0.1:32918");
             discovery.getPeerGroup().setSeedPeerList(seedList);
